requirements.txt
alembic==1.13.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.25
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
